pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bh-aust-postcode"
version = "0.0.1"
description = "Australian Postcode Services."

readme = "README.md"
authors = [{ name = "Van Be Hai Nguyen", email = "behai_nguyen@hotmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["Australia", "Australian", "Postcode"]

dependencies = [
    'tomli; python_version >= "3.10"',
    'Flask', 
    'Flask-RESTX',
    'Flask-Cors',    
    'python-dotenv',
    'requests',    
    'click',
    'simplejson',
    'pytest',
    'coverage',
    'bh-apistatus',
    'bh-utils',
]

requires-python = ">=3.10"

[project.optional-dependencies]
tests = [
  "pytest",
  "coverage"
]
builds = [
    "build", 
    "twine"
]
docs = [
    "sphinx",
    "myst-parser",
    "sphinx-rtd-theme"
]

[project.urls]
repository = "https://github.com/behai-nguyen/bh_aust_postcode"
documentation = "https://bh-aust-postcode.readthedocs.io/"